Updated react-anime
