+++ packages/portfolio/libraries/react-anime/package.json
@@ -1,7 +1,7 @@
 {
   "name": "libraries/react-anime",
   "description": "A super easy animation library for React built on top of anime.js.",
-  "main": "main.tsx",
+  "main": "index.md",
   "keywords": [
     "library",
     "libraries",
